bloom.go

+++ bloom.go
@@ -35,7 +35,7 @@ func (b *BloomFilter) Add(data []byte, p float64, q float64) *BloomFilter {
 			if q==1 && p==0 {
 				r = 0.5
 			} else {
-				r = fastrand.FastRand()/4294967295
+				r = float64(fastrand.FastRand())/float64(4294967295.0)
 			}
 			if j==trueBit {
 				// q chance of returning 1
